# Optimize Bernoulli::new