A small optimization and some clean-up around `Bernoulli::new`. I added some comments to describe the method, and removed the useless "undefined behaviour" test because it does not apply.

Benchmarks before:
```
test misc_gen_bool_const             ... bench:       4,266 ns/iter (+/- 48)
test misc_gen_bool_var               ... bench:       6,239 ns/iter (+/- 63)
test misc_bernoulli_const            ... bench:       3,978 ns/iter (+/- 24)
test misc_bernoulli_var              ... bench:       6,299 ns/iter (+/- 6)
```

After:
```
test misc_gen_bool_const             ... bench:       3,990 ns/iter (+/- 12)
test misc_gen_bool_var               ... bench:       5,984 ns/iter (+/- 90)
test misc_bernoulli_const            ... bench:       3,976 ns/iter (+/- 8)
test misc_bernoulli_var              ... bench:       5,946 ns/iter (+/- 56)
```